requirements.txt
SQLAlchemy==1.1.15
psycopg2==2.7.3.2
pandas==0.21.1
boto3==1.4.4
botocore==1.5.90
docutils==0.13.1
jmespath==0.9.3
python-dateutil==2.6.1
s3transfer==0.1.10
six==1.10.0
xmltodict==0.11.0
awscli==1.14.9
